Fix the header image ratio and crop for mobile devices
Introduce media query for mobile devices.

Introduce three new CSS variables in the .b-hero parent of the header
image. These can be used to adjust the crop of the image to show up.

Introduce auto height and a fixed aspect-ratio for the image in order to
prevent it from being distorted.